refactor(mm-next): move handleGqlResponse
