+++ packages/mirror-media-next/pages/tag/[slug].js
@@ -20,8 +20,11 @@ const GPTAd = dynamic(() => import('../../components/ads/gpt/gpt-ad'), {
 import GPTMbStAd from '../../components/ads/gpt/gpt-mb-st-ad'
 import GPT_Placeholder from '../../components/ads/gpt/gpt-placeholder'
 import { useCallback, useState } from 'react'
-import { getLogTraceObject, handleGqlResponse } from '../../utils'
-import { handleAxiosResponse } from '../../utils/response-handle'
+import { getLogTraceObject } from '../../utils'
+import {
+  handleAxiosResponse,
+  handleGqlResponse,
+} from '../../utils/response-handle'
 
 const TagContainer = styled.main`
   width: 320px;
